+++ src/paintlib/utils/transform-corner.ts
@@ -35,32 +35,57 @@ export class TransformCorner {
     public readonly horizontal: HorizontalPlace,
   ) {}
 
-  getTransformOffset(angle: number, deltaX: number, deltaY: number, scaleX: number, scaleY: number): TBBox {
-    const transform: TBBox = { left: 0, top: 0, width: 0, height: 0 };
+  /**
+   * This function take the actual layout & vector of an object and return the new layout & vector after resizing the
+   * object using given rotation, delta and scale. Used for resizing using object Controls.
+   *
+   * @param layout
+   * @param vector
+   * @param rotation
+   * @param delta
+   * @param scale
+   */
+  transformLayoutVector(layout: TBBox, vector: Point, rotation: number, delta: Point, scale: Point): [TBBox, Point] {
+    layout = { ...layout };
 
     if (this.horizontal === 'l') {
-      // We move the left point: adjust x,y & width
-      transform.width -= deltaX;
+      // We move the left point: adjust x & width
+      layout.width -= delta.x;
 
-      transform.left += Math.cos(angle) * deltaX * scaleX;
-      transform.top += Math.sin(angle) * deltaX * scaleY;
+      layout.left += Math.cos(rotation) * delta.x * scale.x;
+      layout.top += Math.sin(rotation) * delta.x * scale.y;
     } else if (this.horizontal === 'r') {
       // We move the right point: adjust width
-      transform.width += deltaX;
+      layout.width += delta.x;
     }
 
     if (this.vertical === 't') {
       // We move the top point: adjust x,y & height
-      transform.height -= deltaY;
+      layout.height -= delta.y;
 
-      transform.left += Math.cos(angle + Math.PI / 2) * deltaY * scaleX;
-      transform.top += Math.sin(angle + Math.PI / 2) * deltaY * scaleY;
+      layout.left += Math.cos(rotation + Math.PI / 2) * delta.y * scale.x;
+      layout.top += Math.sin(rotation + Math.PI / 2) * delta.y * scale.y;
     } else if (this.vertical === 'b') {
       // We move the bottom point: adjust height
-      transform.height += deltaY;
+      layout.height += delta.y;
     }
 
-    return transform;
+    if (layout.width < 0) {
+      vector = vector.multiply(new Point(-1, 1));
+      layout.width = -layout.width;
+
+      layout.left -= Math.cos(rotation) * layout.width * scale.x;
+      layout.top -= Math.sin(rotation) * layout.width * scale.y;
+    }
+    if (layout.height < 0) {
+      vector = vector.multiply(new Point(1, -1));
+      layout.height = -layout.height;
+
+      layout.left -= Math.cos(rotation + Math.PI / 2) * layout.height * scale.x;
+      layout.top -= Math.sin(rotation + Math.PI / 2) * layout.height * scale.y;
+    }
+
+    return [layout, vector];
   }
 
   toString() {
